revert changes to HTTP roundtripper
